DO NOT MERGE: Size of vector sidExts should be changed to int64_t
Since the size of vector is int, sidExts is determined by a 32-bit variable, but sidExts requires a size of 64 bits to determine "a000000001", and it fails due to a missing digit.
Since the vector size has been fixed in the int64_t in the main branch, the same modification needs to be applied.

Bug: 407442029
Test: VtsHalBroadcastradioAidlTargetTest PerInstance/BroadcastRadioHalTest#DabTune/1_android_hardware_broadcastradio_IBroadcastRadio_dab

+++ broadcastradio/aidl/vts/src/VtsHalBroadcastradioAidlTargetTest.cpp
@@ -753,7 +753,7 @@ TEST_P(BroadcastRadioHalTest, DabTune) {
     LOG(DEBUG) << "Current program info: " << infoCb.toString();
 
     // it should tune exactly to what was requested
-    vector<int> sidExts = bcutils::getAllIds(infoCb.selector, IdentifierType::DAB_SID_EXT);
+    vector<int64_t> sidExts = bcutils::getAllIds(infoCb.selector, IdentifierType::DAB_SID_EXT);
     EXPECT_NE(sidExts.end(), find(sidExts.begin(), sidExts.end(), dabSidExt))
             << "DAB SID ext " << std::hex << dabSidExt << " is not sent back by callback.";
 }
